refactor(memory): extract the observation extraction pipeline to extraction.py
