+++ tests/unit_tests/memory/test_extraction_task.py
@@ -1,87 +1,18 @@
-from contextlib import asynccontextmanager
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from langchain_core.messages import AIMessage, HumanMessage
 from memory.models import MemoryObservation, ObservationStatus
-from memory.schemas import ExtractedObservation, ExtractedObservations
+from memory.schemas import ExtractedObservation
 from memory.tasks import extract_observations_task
-from sessions.models import Run, RunStatus, Session, SessionOrigin
 
-
-def _enabled_config(enabled=True):
-    config = MagicMock()
-    config.memory.enabled = enabled
-    return config
-
-
-def _site_settings(**overrides):
-    """Mock of the site-settings singleton with the memory defaults the task reads."""
-    ss = MagicMock()
-    ss.memory_enabled = True
-    ss.memory_extraction_model_name = "openrouter:openai/gpt-5.4-mini"
-    ss.memory_extraction_fallback_model_name = "openrouter:anthropic/claude-haiku-4.5"
-    for key, value in overrides.items():
-        setattr(ss, key, value)
-    return ss
-
-
-def _checkpointer_with(messages):
-    tup = None
-    if messages is not None:
-        tup = MagicMock()
-        tup.checkpoint = {"channel_values": {"messages": messages}}
-
-    cp = MagicMock()
-    cp.aget_tuple = AsyncMock(return_value=tup)
-
-    @asynccontextmanager
-    async def _open():
-        yield cp
-
-    return _open
-
-
-def _checkpointer_with_delta(writes):
-    """A checkpointer whose ``messages`` is a DeltaChannel: absent from ``channel_values``,
-    recoverable only via ``aget_delta_channel_history``."""
-    tup = MagicMock()
-    tup.checkpoint = {"channel_values": {"session_id": "x"}}  # no messages key
-
-    cp = MagicMock()
-    cp.aget_tuple = AsyncMock(return_value=tup)
-    cp.aget_delta_channel_history = AsyncMock(return_value={"messages": {"writes": writes}})
-
-    @asynccontextmanager
-    async def _open():
-        yield cp
-
-    return _open
-
-
-def _structured_llm_returning(observations=None, *, error=None):
-    llm = MagicMock()
-    if error is not None:
-        llm.with_config.return_value.ainvoke = AsyncMock(side_effect=error)
-    else:
-        llm.with_config.return_value.ainvoke = AsyncMock(
-            return_value=ExtractedObservations(observations=observations or [])
-        )
-    return llm
-
-
-async def _create_run(**kwargs):
-    session = await Session.objects.acreate(
-        thread_id=kwargs.pop("thread_id", "thread-1"),
-        origin=SessionOrigin.API_JOB,
-        repo_id=kwargs.get("repo_id", "group/project"),
-    )
-    defaults = {"trigger_type": SessionOrigin.API_JOB, "repo_id": "group/project", "status": RunStatus.SUCCESSFUL}
-    defaults.update(kwargs)
-    return await Run.objects.acreate(session=session, **defaults)
-
-
-TRANSCRIPT = [HumanMessage(content="fix the bug"), AIMessage(content="done, ran make test")]
+from tests.unit_tests.memory.extraction_helpers import (
+    TRANSCRIPT,
+    _checkpointer_with,
+    _create_run,
+    _enabled_config,
+    _site_settings,
+    _structured_llm_returning,
+)
 
 
 @pytest.mark.django_db(transaction=True)
@@ -95,7 +26,8 @@ async def test_extraction_creates_observation_rows():
     with (
         patch("memory.tasks.RepositoryConfig") as cfg,
         patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm", return_value=_structured_llm_returning(extracted)),
+        patch("memory.extraction.build_structured_llm", return_value=_structured_llm_returning(extracted)),
+        patch("memory.extraction.site_settings", _site_settings()),
     ):
         cfg.get_config.return_value = _enabled_config()
         await extract_observations_task.func(str(run.pk))
@@ -108,151 +40,17 @@ async def test_extraction_creates_observation_rows():
 
 
 @pytest.mark.django_db(transaction=True)
-async def test_extraction_reconstructs_deltachannel_messages():
-    """deepagents' DeltaChannel keeps ``messages`` out of ``channel_values``; extraction
-    must reconstruct the transcript from the delta write history, not skip as empty."""
-    run = await _create_run()
-    writes = [
-        ("t0", "messages", [HumanMessage(content="fix the bug", id="h1")]),
-        ("t1", "messages", [AIMessage(content="done, ran make test", id="a1")]),
-    ]
-    extracted = [ExtractedObservation(category="build_test", content="`make test` sets LANGCHAIN_TRACING_V2=false")]
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with_delta(writes)),
-        patch("memory.tasks._build_structured_llm", return_value=_structured_llm_returning(extracted)),
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))
-
-    rows = [obs async for obs in MemoryObservation.objects.filter(repo_id="group/project")]
-    assert len(rows) == 1
-    assert rows[0].category == "build_test"
-
-
-@pytest.mark.django_db(transaction=True)
-async def test_extraction_skips_when_checkpoint_expired():
-    run = await _create_run()
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with(None)),
-        patch("memory.tasks._build_structured_llm") as build,
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))  # must not raise
-
-    build.assert_not_called()
-    assert await MemoryObservation.objects.acount() == 0
-
-
-@pytest.mark.django_db(transaction=True)
-async def test_extraction_warns_when_checkpoint_has_no_messages(caplog):
-    # A present checkpoint with an empty message list is a defect signature, distinct from
-    # a missing/expired checkpoint: it skips like the expired case but logs at WARNING.
-    run = await _create_run()
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with([])),
-        patch("memory.tasks._build_structured_llm") as build,
-        caplog.at_level("WARNING", logger="daiv.memory"),
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))  # must not raise
-
-    build.assert_not_called()
-    assert await MemoryObservation.objects.acount() == 0
-    assert any("has no messages" in record.message for record in caplog.records)
-
-
-@pytest.mark.django_db(transaction=True)
-async def test_extraction_skips_transcript_without_ai_turns(caplog):
-    # A run that died before the agent loop started (sandbox unavailable, auth failure) can only
-    # produce an empty extraction — the model call is pure cost, so it is never made.
-    run = await _create_run(status=RunStatus.FAILED)
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with([HumanMessage(content="fix the bug")])),
-        patch("memory.tasks._build_structured_llm") as build,
-        caplog.at_level("INFO", logger="daiv.memory"),
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))
-
-    build.assert_not_called()
-    assert await MemoryObservation.objects.acount() == 0
-    assert any("no AI turns" in record.message for record in caplog.records)
-
-
-@pytest.mark.django_db(transaction=True)
 async def test_extraction_respects_daiv_yml_flag():
     run = await _create_run()
 
     with (
         patch("memory.tasks.RepositoryConfig") as cfg,
         patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm") as build,
+        patch("memory.extraction.build_structured_llm") as build,
+        patch("memory.extraction.site_settings", _site_settings()),
     ):
         cfg.get_config.return_value = _enabled_config(enabled=False)
         await extract_observations_task.func(str(run.pk))
-
-    build.assert_not_called()
-    assert await MemoryObservation.objects.acount() == 0
-
-
-@pytest.mark.django_db(transaction=True)
-@pytest.mark.parametrize(
-    ("fallback_model", "expected_models"),
-    [
-        # Both configured → both passed through.
-        ("provider:m2", ("provider:m1", "provider:m2")),
-        # Empty fallback is filtered out so _build_structured_llm gets a 1-tuple, not (model, None).
-        (None, ("provider:m1",)),
-    ],
-    ids=["with_fallback", "drops_empty_fallback"],
-)
-async def test_extraction_uses_configured_models(fallback_model, expected_models):
-    run = await _create_run()
-    extracted = [ExtractedObservation(category="build_test", content="`make test` needs the DB up first")]
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm", return_value=_structured_llm_returning(extracted)) as build,
-        patch(
-            "memory.tasks.site_settings",
-            _site_settings(
-                memory_extraction_model_name="provider:m1", memory_extraction_fallback_model_name=fallback_model
-            ),
-        ),
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))
-
-    _schema, models = build.call_args.args
-    assert tuple(models) == expected_models
-
-
-@pytest.mark.django_db(transaction=True)
-async def test_extraction_noop_when_no_model_configured():
-    # Both model and fallback empty (only reachable via an empty-string env override) → clean skip,
-    # not an IndexError crash in _build_structured_llm.
-    run = await _create_run()
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm") as build,
-        patch(
-            "memory.tasks.site_settings",
-            _site_settings(memory_extraction_model_name="", memory_extraction_fallback_model_name=""),
-        ),
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))  # must not raise
 
     build.assert_not_called()
     assert await MemoryObservation.objects.acount() == 0
@@ -266,7 +64,7 @@ async def test_extraction_noop_when_site_disabled():
     with (
         patch("memory.tasks.RepositoryConfig") as cfg,
         patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm") as build,
+        patch("memory.extraction.build_structured_llm") as build,
         patch("memory.tasks.site_settings", _site_settings(memory_enabled=False)),
     ):
         cfg.get_config.return_value = _enabled_config(enabled=True)
@@ -295,30 +93,12 @@ async def test_extraction_skips_run_without_session_id():
         mock_qs.afirst = AsyncMock(return_value=run)
         mock_run.objects.filter.return_value = mock_qs
 
-        with patch("memory.tasks.RepositoryConfig") as cfg, patch("memory.tasks._build_structured_llm") as build:
+        with patch("memory.tasks.RepositoryConfig") as cfg, patch("memory.extraction.build_structured_llm") as build:
             await extract_observations_task.func(str(run.pk))  # must not raise
 
         cfg.get_config.assert_not_called()  # bails before loading config
         build.assert_not_called()
         assert await MemoryObservation.objects.acount() == 0
-
-
-@pytest.mark.django_db(transaction=True)
-async def test_extraction_noop_when_model_spec_invalid():
-    # A bad/unparseable extraction model spec raises ValueError; it must be swallowed (clean skip),
-    # not crash the task. The hardcoded extraction models raise this in a deployment without the
-    # OpenAI/Anthropic provider rows configured (regression guard for C1).
-    run = await _create_run()
-
-    with (
-        patch("memory.tasks.RepositoryConfig") as cfg,
-        patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm", side_effect=ValueError("Unknown/Unsupported provider for model")),
-    ):
-        cfg.get_config.return_value = _enabled_config()
-        await extract_observations_task.func(str(run.pk))  # must not raise
-
-    assert await MemoryObservation.objects.acount() == 0
 
 
 @pytest.mark.django_db(transaction=True)
@@ -332,7 +112,8 @@ async def test_extraction_propagates_llm_failure_without_partial_writes():
     with (
         patch("memory.tasks.RepositoryConfig") as cfg,
         patch("core.checkpointer.open_checkpointer", _checkpointer_with(TRANSCRIPT)),
-        patch("memory.tasks._build_structured_llm", return_value=failing_llm),
+        patch("memory.extraction.build_structured_llm", return_value=failing_llm),
+        patch("memory.extraction.site_settings", _site_settings()),
         pytest.raises(RuntimeError),
     ):
         cfg.get_config.return_value = _enabled_config()
